requirements.txt
langchain
langchain-ollama
langchain-community
langchain-text-splitters
chromadb
pypdf
python-dotenv
langchain-chroma